+++ pkg/engine/plan/analyze_plan_kind_test.go
@@ -1,4 +1,4 @@
-package planv2
+package plan
 
 import (
 	"testing"
