+++ pkg/postprocess/stream_test.go
@@ -17,81 +17,65 @@ func TestProcessStream_Process(t *testing.T) {
 
 	original := &plan.SynchronousResponsePlan{
 		FlushInterval: 500,
-		Response: resolve.GraphQLResponse{
+		Response: &resolve.GraphQLResponse{
 			Data: &resolve.Object{
 				Fetch: &resolve.SingleFetch{
 					DataSource: userService,
 					BufferId:   0,
 				},
-				FieldSets: []resolve.FieldSet{
+				Fields: []*resolve.Field{
 					{
 						HasBuffer: true,
 						BufferID:  0,
-						Fields: []resolve.Field{
-							{
-								Name: []byte("users"),
-								Stream: &resolve.StreamField{
-									InitialBatchSize: 0,
-								},
-								Value: &resolve.Array{
-									Item: &resolve.Object{
-										Fetch: &resolve.SingleFetch{
-											BufferId:   1,
-											DataSource: postsService,
-											InputTemplate: resolve.InputTemplate{
-												Segments: []resolve.TemplateSegment{
-													{
-														SegmentType:        resolve.VariableSegmentType,
-														VariableSource:     resolve.VariableSourceObject,
-														VariableSourcePath: []string{"id"},
-													},
-												},
+						Name:      []byte("users"),
+						Stream: &resolve.StreamField{
+							InitialBatchSize: 0,
+						},
+						Value: &resolve.Array{
+							Item: &resolve.Object{
+								Fetch: &resolve.SingleFetch{
+									BufferId:   1,
+									DataSource: postsService,
+									InputTemplate: resolve.InputTemplate{
+										Segments: []resolve.TemplateSegment{
+											{
+												SegmentType:        resolve.VariableSegmentType,
+												VariableSource:     resolve.VariableSourceObject,
+												VariableSourcePath: []string{"id"},
 											},
 										},
-										FieldSets: []resolve.FieldSet{
-											{
-												Fields: []resolve.Field{
+									},
+								},
+								Fields: []*resolve.Field{
+									{
+										Name: []byte("id"),
+										Value: &resolve.Integer{
+											Path: []string{"id"},
+										},
+									},
+									{
+										Name: []byte("name"),
+										Value: &resolve.String{
+											Path: []string{"name"},
+										},
+									},
+									{
+										HasBuffer: true,
+										BufferID:  1,
+										Name:      []byte("posts"),
+										Value: &resolve.Array{
+											Item: &resolve.Object{
+												Fields: []*resolve.Field{
 													{
-														Name: []byte("id"),
-														Value: &resolve.Integer{
-															Path: []string{"id"},
-														},
-													},
-													{
-														Name: []byte("name"),
+														Name: []byte("title"),
 														Value: &resolve.String{
-															Path: []string{"name"},
+															Path: []string{"title"},
 														},
 													},
-												},
-											},
-											{
-												HasBuffer: true,
-												BufferID:  1,
-												Fields: []resolve.Field{
 													{
-														Name:  []byte("posts"),
-														Value: &resolve.Array{
-															Item: &resolve.Object{
-																FieldSets: []resolve.FieldSet{
-																	{
-																		Fields: []resolve.Field{
-																			{
-																				Name: []byte("title"),
-																				Value: &resolve.String{
-																					Path: []string{"title"},
-																				},
-																			},
-																			{
-																				Name: []byte("body"),
-																				Value: &resolve.String{
-																					Path: []string{"body"},
-																				},
-																			},
-																		},
-																	},
-																},
-															},
+														Name: []byte("body"),
+														Value: &resolve.String{
+															Path: []string{"body"},
 														},
 													},
 												},
@@ -117,20 +101,16 @@ func TestProcessStream_Process(t *testing.T) {
 						DataSource: userService,
 						BufferId:   0,
 					},
-					FieldSets: []resolve.FieldSet{
+					Fields: []*resolve.Field{
 						{
 							HasBuffer: true,
 							BufferID:  0,
-							Fields: []resolve.Field{
-								{
-									Name: []byte("users"),
-									Value: &resolve.Array{
-										Stream: resolve.Stream{
-											Enabled: true,
-											InitialBatchSize: 0,
-											PatchIndex: 0,
-										},
-									},
+							Name:      []byte("users"),
+							Value: &resolve.Array{
+								Stream: resolve.Stream{
+									Enabled:          true,
+									InitialBatchSize: 0,
+									PatchIndex:       0,
 								},
 							},
 						},
@@ -154,48 +134,37 @@ func TestProcessStream_Process(t *testing.T) {
 								},
 							},
 						},
-						FieldSets: []resolve.FieldSet{
+						Fields: []*resolve.Field{
 							{
-								Fields: []resolve.Field{
-									{
-										Name: []byte("id"),
-										Value: &resolve.Integer{
-											Path: []string{"id"},
-										},
-									},
-									{
-										Name: []byte("name"),
-										Value: &resolve.String{
-											Path: []string{"name"},
-										},
-									},
+								Name: []byte("id"),
+								Value: &resolve.Integer{
+									Path: []string{"id"},
 								},
 							},
 							{
+								Name: []byte("name"),
+								Value: &resolve.String{
+									Path: []string{"name"},
+								},
+							},
+							{
+
 								HasBuffer: true,
 								BufferID:  1,
-								Fields: []resolve.Field{
-									{
-										Name:  []byte("posts"),
-										Value: &resolve.Array{
-											Item: &resolve.Object{
-												FieldSets: []resolve.FieldSet{
-													{
-														Fields: []resolve.Field{
-															{
-																Name: []byte("title"),
-																Value: &resolve.String{
-																	Path: []string{"title"},
-																},
-															},
-															{
-																Name: []byte("body"),
-																Value: &resolve.String{
-																	Path: []string{"body"},
-																},
-															},
-														},
-													},
+								Name:      []byte("posts"),
+								Value: &resolve.Array{
+									Item: &resolve.Object{
+										Fields: []*resolve.Field{
+											{
+												Name: []byte("title"),
+												Value: &resolve.String{
+													Path: []string{"title"},
+												},
+											},
+											{
+												Name: []byte("body"),
+												Value: &resolve.String{
+													Path: []string{"body"},
 												},
 											},
 										},
@@ -222,81 +191,65 @@ func TestProcessStream_Process_BatchSize_1(t *testing.T) {
 
 	original := &plan.SynchronousResponsePlan{
 		FlushInterval: 500,
-		Response: resolve.GraphQLResponse{
+		Response: &resolve.GraphQLResponse{
 			Data: &resolve.Object{
 				Fetch: &resolve.SingleFetch{
 					DataSource: userService,
 					BufferId:   0,
 				},
-				FieldSets: []resolve.FieldSet{
+				Fields: []*resolve.Field{
 					{
 						HasBuffer: true,
 						BufferID:  0,
-						Fields: []resolve.Field{
-							{
-								Name: []byte("users"),
-								Stream: &resolve.StreamField{
-									InitialBatchSize: 1,
-								},
-								Value: &resolve.Array{
-									Item: &resolve.Object{
-										Fetch: &resolve.SingleFetch{
-											BufferId:   1,
-											DataSource: postsService,
-											InputTemplate: resolve.InputTemplate{
-												Segments: []resolve.TemplateSegment{
-													{
-														SegmentType:        resolve.VariableSegmentType,
-														VariableSource:     resolve.VariableSourceObject,
-														VariableSourcePath: []string{"id"},
-													},
-												},
+						Name:      []byte("users"),
+						Stream: &resolve.StreamField{
+							InitialBatchSize: 1,
+						},
+						Value: &resolve.Array{
+							Item: &resolve.Object{
+								Fetch: &resolve.SingleFetch{
+									BufferId:   1,
+									DataSource: postsService,
+									InputTemplate: resolve.InputTemplate{
+										Segments: []resolve.TemplateSegment{
+											{
+												SegmentType:        resolve.VariableSegmentType,
+												VariableSource:     resolve.VariableSourceObject,
+												VariableSourcePath: []string{"id"},
 											},
 										},
-										FieldSets: []resolve.FieldSet{
-											{
-												Fields: []resolve.Field{
+									},
+								},
+								Fields: []*resolve.Field{
+									{
+										Name: []byte("id"),
+										Value: &resolve.Integer{
+											Path: []string{"id"},
+										},
+									},
+									{
+										Name: []byte("name"),
+										Value: &resolve.String{
+											Path: []string{"name"},
+										},
+									},
+									{
+										HasBuffer: true,
+										BufferID:  1,
+										Name:      []byte("posts"),
+										Value: &resolve.Array{
+											Item: &resolve.Object{
+												Fields: []*resolve.Field{
 													{
-														Name: []byte("id"),
-														Value: &resolve.Integer{
-															Path: []string{"id"},
-														},
-													},
-													{
-														Name: []byte("name"),
+														Name: []byte("title"),
 														Value: &resolve.String{
-															Path: []string{"name"},
+															Path: []string{"title"},
 														},
 													},
-												},
-											},
-											{
-												HasBuffer: true,
-												BufferID:  1,
-												Fields: []resolve.Field{
 													{
-														Name:  []byte("posts"),
-														Value: &resolve.Array{
-															Item: &resolve.Object{
-																FieldSets: []resolve.FieldSet{
-																	{
-																		Fields: []resolve.Field{
-																			{
-																				Name: []byte("title"),
-																				Value: &resolve.String{
-																					Path: []string{"title"},
-																				},
-																			},
-																			{
-																				Name: []byte("body"),
-																				Value: &resolve.String{
-																					Path: []string{"body"},
-																				},
-																			},
-																		},
-																	},
-																},
-															},
+														Name: []byte("body"),
+														Value: &resolve.String{
+															Path: []string{"body"},
 														},
 													},
 												},
@@ -322,76 +275,62 @@ func TestProcessStream_Process_BatchSize_1(t *testing.T) {
 						DataSource: userService,
 						BufferId:   0,
 					},
-					FieldSets: []resolve.FieldSet{
+					Fields: []*resolve.Field{
 						{
 							HasBuffer: true,
 							BufferID:  0,
-							Fields: []resolve.Field{
-								{
-									Name: []byte("users"),
-									Value: &resolve.Array{
-										Stream: resolve.Stream{
-											Enabled: true,
-											InitialBatchSize: 1,
-										},
-										Item: &resolve.Object{
-											Fetch: &resolve.SingleFetch{
-												BufferId:   1,
-												DataSource: postsService,
-												InputTemplate: resolve.InputTemplate{
-													Segments: []resolve.TemplateSegment{
-														{
-															SegmentType:        resolve.VariableSegmentType,
-															VariableSource:     resolve.VariableSourceObject,
-															VariableSourcePath: []string{"id"},
-														},
-													},
+							Name:      []byte("users"),
+							Value: &resolve.Array{
+								Stream: resolve.Stream{
+									Enabled:          true,
+									InitialBatchSize: 1,
+								},
+								Item: &resolve.Object{
+									Fetch: &resolve.SingleFetch{
+										BufferId:   1,
+										DataSource: postsService,
+										InputTemplate: resolve.InputTemplate{
+											Segments: []resolve.TemplateSegment{
+												{
+													SegmentType:        resolve.VariableSegmentType,
+													VariableSource:     resolve.VariableSourceObject,
+													VariableSourcePath: []string{"id"},
 												},
 											},
-											FieldSets: []resolve.FieldSet{
-												{
-													Fields: []resolve.Field{
+										},
+									},
+									Fields: []*resolve.Field{
+										{
+											Name: []byte("id"),
+											Value: &resolve.Integer{
+												Path: []string{"id"},
+											},
+										},
+										{
+											Name: []byte("name"),
+											Value: &resolve.String{
+												Path: []string{"name"},
+											},
+										},
+
+										{
+
+											HasBuffer: true,
+											BufferID:  1,
+											Name:      []byte("posts"),
+											Value: &resolve.Array{
+												Item: &resolve.Object{
+													Fields: []*resolve.Field{
 														{
-															Name: []byte("id"),
-															Value: &resolve.Integer{
-																Path: []string{"id"},
-															},
-														},
-														{
-															Name: []byte("name"),
+															Name: []byte("title"),
 															Value: &resolve.String{
-																Path: []string{"name"},
+																Path: []string{"title"},
 															},
 														},
-													},
-												},
-												{
-													HasBuffer: true,
-													BufferID:  1,
-													Fields: []resolve.Field{
 														{
-															Name:  []byte("posts"),
-															Value: &resolve.Array{
-																Item: &resolve.Object{
-																	FieldSets: []resolve.FieldSet{
-																		{
-																			Fields: []resolve.Field{
-																				{
-																					Name: []byte("title"),
-																					Value: &resolve.String{
-																						Path: []string{"title"},
-																					},
-																				},
-																				{
-																					Name: []byte("body"),
-																					Value: &resolve.String{
-																						Path: []string{"body"},
-																					},
-																				},
-																			},
-																		},
-																	},
-																},
+															Name: []byte("body"),
+															Value: &resolve.String{
+																Path: []string{"body"},
 															},
 														},
 													},
@@ -422,48 +361,36 @@ func TestProcessStream_Process_BatchSize_1(t *testing.T) {
 								},
 							},
 						},
-						FieldSets: []resolve.FieldSet{
+						Fields: []*resolve.Field{
 							{
-								Fields: []resolve.Field{
-									{
-										Name: []byte("id"),
-										Value: &resolve.Integer{
-											Path: []string{"id"},
-										},
-									},
-									{
-										Name: []byte("name"),
-										Value: &resolve.String{
-											Path: []string{"name"},
-										},
-									},
+								Name: []byte("id"),
+								Value: &resolve.Integer{
+									Path: []string{"id"},
+								},
+							},
+							{
+								Name: []byte("name"),
+								Value: &resolve.String{
+									Path: []string{"name"},
 								},
 							},
 							{
 								HasBuffer: true,
 								BufferID:  1,
-								Fields: []resolve.Field{
-									{
-										Name:  []byte("posts"),
-										Value: &resolve.Array{
-											Item: &resolve.Object{
-												FieldSets: []resolve.FieldSet{
-													{
-														Fields: []resolve.Field{
-															{
-																Name: []byte("title"),
-																Value: &resolve.String{
-																	Path: []string{"title"},
-																},
-															},
-															{
-																Name: []byte("body"),
-																Value: &resolve.String{
-																	Path: []string{"body"},
-																},
-															},
-														},
-													},
+								Name:      []byte("posts"),
+								Value: &resolve.Array{
+									Item: &resolve.Object{
+										Fields: []*resolve.Field{
+											{
+												Name: []byte("title"),
+												Value: &resolve.String{
+													Path: []string{"title"},
+												},
+											},
+											{
+												Name: []byte("body"),
+												Value: &resolve.String{
+													Path: []string{"body"},
 												},
 											},
 										},
